refactor: rename DTsearchUtils to DTSearchQueryUtils
